chore: Refactor load_pointcloudset function to support axis inversion
The load_pointcloudset function in get_pointcloud_data.py has been refactored to support axis inversion. This allows the user to specify a list of axes to invert (multiply by -1) when loading the point cloud dataset from a ROS bag file. The function now takes an additional optional parameter, invert_axes, which accepts a list of axes to invert. If invert_axes is provided, the function applies the inversion to the dataset before returning it.